fix(ci): stop wrapper.boundary.test.ts flaking on Windows pkg downloads
The Claude wrapper boundary test compiles a fake claude.exe with
@yao-pkg/pkg on Windows; on a cold pkg cache that first downloads a base
Node binary from GitHub inside vitest's 10s beforeAll budget — the
dominant Windows CI flake. Two composing fixes, no retries:

- Cache PKG_CACHE_PATH in the setup action (Windows only), keyed on
  OS/arch + host Node major + pkg version, restore-always/save-on-main
  like the pnpm store. Since pkg only populates the cache at test time,
  a main-gated warmup compile seeds it before saving.
- Compile the fake binary in a boundary-project globalSetup (no hook
  timeout applies) and hand the path to tests via provide/inject, so
  even a cold cache cannot time out a test hook.
