+++ src/modules/agent-module/claude/wrapper.boundary.test.ts
@@ -11,7 +11,7 @@
  * - Exit code propagation
  */
 
-import { describe, it, expect, beforeEach, afterEach, beforeAll, afterAll } from "vitest";
+import { describe, it, expect, beforeEach, afterEach, beforeAll, inject } from "vitest";
 import { join, resolve, dirname, delimiter } from "node:path";
 import { writeFile, mkdir } from "node:fs/promises";
 import { existsSync } from "node:fs";
@@ -70,62 +70,6 @@ function parseAllFakeClaudeOutputs(stdout: string): FakeClaudeOutput[] {
 }
 
 /**
- * Create a fake claude binary for testing.
- * Outputs JSON with received args and selected env vars, exits with configurable code.
- *
- * Supports per-invocation exit codes via CLAUDE_EXIT_CODES env var (comma-separated)
- * and CLAUDE_COUNTER_FILE for tracking invocation count across calls.
- */
-async function createFakeClaudeBinary(binDir: string): Promise<string> {
-  const fakeNodeContent = `#!/usr/bin/env node
-const fs = require("node:fs");
-
-// Always succeed for --version (used by findSystemClaude discovery)
-if (process.argv.includes("--version")) {
-  console.log("fake-claude 1.0.0");
-  process.exit(0);
-}
-
-const output = {
-  args: process.argv.slice(2),
-  env: {
-    CLAUDECODE: process.env.CLAUDECODE ?? null,
-  },
-};
-
-// Track invocation count for multi-call tests
-const counterFile = process.env.CLAUDE_COUNTER_FILE;
-let callIndex = 0;
-if (counterFile) {
-  try {
-    callIndex = parseInt(fs.readFileSync(counterFile, "utf-8"), 10);
-  } catch { /* first call */ }
-  fs.writeFileSync(counterFile, String(callIndex + 1));
-}
-
-// Support per-invocation exit codes: "1,0" means first exits 1, second exits 0
-const exitCodes = process.env.CLAUDE_EXIT_CODES;
-let exitCode = 0;
-if (exitCodes) {
-  const codes = exitCodes.split(",").map(Number);
-  exitCode = codes[callIndex] ?? codes[codes.length - 1] ?? 0;
-} else {
-  exitCode = parseInt(process.env.CLAUDE_EXIT_CODE || "0", 10);
-}
-
-console.log(JSON.stringify(output));
-process.exit(isNaN(exitCode) ? 0 : exitCode);
-`;
-  // Real .exe on Windows so shell:false works (no cmd.exe involvement)
-  return createFakeAgentBinary({
-    dir: binDir,
-    binaryName: "claude",
-    scriptBody: fakeNodeContent,
-    windowsMode: "exe",
-  });
-}
-
-/**
  * Build a PATH that includes the fake binary dir and node's directory.
  */
 function buildPath(fakeBinDir: string): string {
@@ -134,23 +78,18 @@ function buildPath(fakeBinDir: string): string {
 
 describe("ch-claude.cjs boundary tests", () => {
   let tempDir: { path: string; cleanup: () => Promise<void> };
-  let sharedBinTempDir: { path: string; cleanup: () => Promise<void> };
   let fakeBinDir: string;
 
   beforeAll(async () => {
     await assertCompiledScript(COMPILED_SCRIPT_PATH);
-    // Compile the fake claude binary ONCE for the whole file. On Windows,
-    // createFakeClaudeBinary invokes @yao-pkg/pkg to bundle a real .exe, which
-    // takes seconds; doing it per-test in beforeEach blew the default 10s hook
-    // timeout under CI contention (the dominant flaky-test failure). The fake
+    // The fake claude binary is compiled once per run by the boundary
+    // project's globalSetup (src/test/global-setup-boundary.ts). On Windows
+    // that means bundling a real .exe with @yao-pkg/pkg — on a cold pkg cache
+    // a network download plus a multi-second compile, which blew the 10s hook
+    // timeout when it ran here (the dominant flaky-test failure). The fake
     // binary's behavior is driven entirely by env vars (CLAUDE_EXIT_CODE,
     // CLAUDE_COUNTER_FILE, …), so a single shared binary serves every test.
-    sharedBinTempDir = await createTempDir();
-    fakeBinDir = await createFakeClaudeBinary(join(sharedBinTempDir.path, "bin"));
-  });
-
-  afterAll(async () => {
-    await sharedBinTempDir.cleanup();
+    fakeBinDir = inject("fakeClaudeBinDir");
   });
 
   beforeEach(async () => {
